modify double to long

+++ src/main/java/com/linkedin/kmf/services/ConsumeService.java
@@ -173,7 +173,7 @@ public class ConsumeService implements Service {
         _sensors._recordsDuplicated.record();
       } else if (index > nextIndex) {
         nextIndexes.put(partition, index + 1);
-        double numLostRecords = index - nextIndex;
+        long numLostRecords = index - nextIndex;
         _sensors._recordsLost.record(numLostRecords);
         LOG.info("_recordsLost recorded: Avro record current index: {} at {}. Next index: {}. Lost {} records.", index, currMs, nextIndex, numLostRecords);
       }
